requirements.txt
asgiref==3.8.1
blessed==1.20.0
boto3==1.34.144
botocore==1.34.150
cement==2.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.3
Django==3.2.25
django-cors-headers==4.3.1
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.16
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pathspec==0.10.1
pdfrw==0.4
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.1.9
